split browser module into two files
