+++ wgpu-renderer/src/browser/state.rs
@@ -1,4 +1,3 @@
-use crate::args::Args;
 use wgpu::TextureFormat;
 use wgpu_text::{
     font::FontArc,
@@ -6,68 +5,6 @@ use wgpu_text::{
     BrushBuilder, TextBrush,
 };
 use winit::{event::WindowEvent, window::Window};
-use winit::{
-    event::*,
-    event_loop::{ControlFlow, EventLoop},
-    window::WindowBuilder,
-};
-
-pub async fn run(args: Args) {
-    let event_loop = EventLoop::new();
-    let window = WindowBuilder::new()
-        .with_title("benser")
-        .build(&event_loop)
-        .unwrap();
-
-    let mut state = State::new(window).await;
-
-    event_loop.run(move |event, _, control_flow| match event {
-        Event::RedrawRequested(window_id) if window_id == state.window().id() => {
-            state.update();
-            match state.render() {
-                Ok(_) => {}
-                // Reconfigure the surface if lost
-                Err(wgpu::SurfaceError::Lost) => state.resize(state.size),
-                // The system is out of memory, we should probably quit
-                Err(wgpu::SurfaceError::OutOfMemory) => *control_flow = ControlFlow::Exit,
-                // All other errors (Outdated, Timeout) should be resolved by the next frame
-                Err(e) => eprintln!("{:?}", e),
-            }
-        }
-        Event::MainEventsCleared => {
-            // RedrawRequested will only trigger once, unless we manually
-            // request it.
-            state.window().request_redraw();
-        }
-        Event::WindowEvent {
-            ref event,
-            window_id,
-        } if window_id == state.window().id() => {
-            if !state.input(event) {
-                match event {
-                    WindowEvent::Resized(physical_size) => {
-                        state.resize(*physical_size);
-                    }
-                    WindowEvent::ScaleFactorChanged { new_inner_size, .. } => {
-                        state.resize(**new_inner_size);
-                    }
-                    WindowEvent::CloseRequested
-                    | WindowEvent::KeyboardInput {
-                        input:
-                            KeyboardInput {
-                                state: ElementState::Pressed,
-                                virtual_keycode: Some(VirtualKeyCode::Escape),
-                                ..
-                            },
-                        ..
-                    } => *control_flow = ControlFlow::Exit,
-                    _ => {}
-                }
-            }
-        }
-        _ => {}
-    });
-}
 
 pub struct State {
     window: Window,
@@ -136,7 +73,7 @@ impl State {
         };
         surface.configure(&device, &config);
 
-        let font = FontArc::try_from_slice(include_bytes!("fonts/OpenSans.ttf")).unwrap();
+        let font = FontArc::try_from_slice(include_bytes!("../fonts/OpenSans.ttf")).unwrap();
 
         let brush = BrushBuilder::using_font(font).build(
             &device,
